Resolve promise to avoid hang when handling files
Returning inside a Promise does not resolve said Promise. The calling code will hang forever. This can easily be demostrated by using die virtual adapter with diretories and with and without a file entry. As soon as there is a file entry, it hangs.
Resolving the promise solves this bug.